HBAI-77 Move code into modules
