--- backends/inference/text_routes.py
@@ -1,6 +1,0 @@
-def inference_completions(query: str) -> str:
-    # @TODO Load the normal text llm completion..
-    # Or use llamaIndex's abstraction
-    # llm.stream_complete()
-
-    return query
